+++ RELEASES.md
@@ -7,6 +7,8 @@ Upcoming
 --------
 
 - Refactor XmlStorage's MetadataHandler in order to minimize rewrites.
+- Fix change in report name after replacement for XmlStorage
+- Add thread with warning when threadCreatepoint() (and threadStartpoint()) not used
 - Fix metadata file format for XmlStorage
 - Improve storage id handling of XmlStorage while reading from a filesystem.
 - Fix report discovery for XmlStorage